Flatten the expecting directory hierarchy a bit.
Mostly just shorter names. Originally the hr-st repo expected to
commit rendered outputs to the repo to simplify access, but this
made the repo very large, so we switched to auto-generating and
deploying the site to a public location.

That makes it less important to keep separate directories for all
the temporary files, and so we can use less-specific naming for
their directories.