Click to keep a line of hieroglyphs visible.
Using the display-on-fade implementation last session, we ran into
an issue with wanting the heiroglyphic transcription visible for
reference sometimes when the mouse was focussed on annotation
elsewhere. Switch to a `fade` class attribute and add a click
handler to turn it on and off.

With this is should be possible to click on a transcription and
have it stay visible afterward.